Move rustc and target configs from BuildConfig into BuildContext
