+++ src/cargo/core/compiler/build_config.rs
@@ -1,18 +1,10 @@
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-use util::{CargoResult, CargoResultExt, Config, Rustc};
-use core::Workspace;
-use super::BuildOutput;
+use std::path::Path;
+use util::{CargoResult, CargoResultExt, Config};
 
 /// Configuration information for a rustc build.
 pub struct BuildConfig {
-    pub rustc: Rustc,
-    /// Build information for the host arch
-    pub host: TargetConfig,
     /// The target arch triple, defaults to host arch
     pub requested_target: Option<String>,
-    /// Build information for the target
-    pub target: TargetConfig,
     /// How many rustc jobs to run in parallel
     pub jobs: u32,
     /// Whether we are building for release
@@ -36,7 +28,6 @@ impl BuildConfig {
         config: &Config,
         jobs: Option<u32>,
         requested_target: &Option<String>,
-        ws: Option<&Workspace>,
         mode: CompileMode,
     ) -> CargoResult<BuildConfig> {
         let requested_target = match requested_target {
@@ -89,39 +80,13 @@ impl BuildConfig {
             None => None,
         };
         let jobs = jobs.or(cfg_jobs).unwrap_or(::num_cpus::get() as u32);
-        let rustc = config.rustc(ws)?;
-        let host_config = TargetConfig::new(config, &rustc.host)?;
-        let target_config = match target.as_ref() {
-            Some(triple) => TargetConfig::new(config, triple)?,
-            None => host_config.clone(),
-        };
         Ok(BuildConfig {
-            rustc,
             requested_target: target,
             jobs,
-            host: host_config,
-            target: target_config,
             release: false,
             mode,
             message_format: MessageFormat::Human,
         })
-    }
-
-    /// The host arch triple
-    ///
-    /// e.g. x86_64-unknown-linux-gnu, would be
-    ///  - machine: x86_64
-    ///  - hardware-platform: unknown
-    ///  - operating system: linux-gnu
-    pub fn host_triple(&self) -> &str {
-        &self.rustc.host
-    }
-
-    pub fn target_triple(&self) -> &str {
-        self.requested_target
-            .as_ref()
-            .map(|s| s.as_str())
-            .unwrap_or(self.host_triple())
     }
 
     pub fn json_messages(&self) -> bool {
@@ -220,96 +185,5 @@ impl CompileMode {
             CompileMode::RunCustomBuild,
         ];
         &ALL
-    }
-}
-
-/// Information required to build for a target
-#[derive(Clone, Default)]
-pub struct TargetConfig {
-    /// The path of archiver (lib builder) for this target.
-    pub ar: Option<PathBuf>,
-    /// The path of the linker for this target.
-    pub linker: Option<PathBuf>,
-    /// Special build options for any necessary input files (filename -> options)
-    pub overrides: HashMap<String, BuildOutput>,
-}
-
-impl TargetConfig {
-    pub fn new(config: &Config, triple: &str) -> CargoResult<TargetConfig> {
-        let key = format!("target.{}", triple);
-        let mut ret = TargetConfig {
-            ar: config.get_path(&format!("{}.ar", key))?.map(|v| v.val),
-            linker: config.get_path(&format!("{}.linker", key))?.map(|v| v.val),
-            overrides: HashMap::new(),
-        };
-        let table = match config.get_table(&key)? {
-            Some(table) => table.val,
-            None => return Ok(ret),
-        };
-        for (lib_name, value) in table {
-            match lib_name.as_str() {
-                "ar" | "linker" | "runner" | "rustflags" => continue,
-                _ => {}
-            }
-
-            let mut output = BuildOutput {
-                library_paths: Vec::new(),
-                library_links: Vec::new(),
-                cfgs: Vec::new(),
-                env: Vec::new(),
-                metadata: Vec::new(),
-                rerun_if_changed: Vec::new(),
-                rerun_if_env_changed: Vec::new(),
-                warnings: Vec::new(),
-            };
-            // We require deterministic order of evaluation, so we must sort the pairs by key first.
-            let mut pairs = Vec::new();
-            for (k, value) in value.table(&lib_name)?.0 {
-                pairs.push((k, value));
-            }
-            pairs.sort_by_key(|p| p.0);
-            for (k, value) in pairs {
-                let key = format!("{}.{}", key, k);
-                match &k[..] {
-                    "rustc-flags" => {
-                        let (flags, definition) = value.string(k)?;
-                        let whence = format!("in `{}` (in {})", key, definition.display());
-                        let (paths, links) = BuildOutput::parse_rustc_flags(flags, &whence)?;
-                        output.library_paths.extend(paths);
-                        output.library_links.extend(links);
-                    }
-                    "rustc-link-lib" => {
-                        let list = value.list(k)?;
-                        output
-                            .library_links
-                            .extend(list.iter().map(|v| v.0.clone()));
-                    }
-                    "rustc-link-search" => {
-                        let list = value.list(k)?;
-                        output
-                            .library_paths
-                            .extend(list.iter().map(|v| PathBuf::from(&v.0)));
-                    }
-                    "rustc-cfg" => {
-                        let list = value.list(k)?;
-                        output.cfgs.extend(list.iter().map(|v| v.0.clone()));
-                    }
-                    "rustc-env" => for (name, val) in value.table(k)?.0 {
-                        let val = val.string(name)?.0;
-                        output.env.push((name.clone(), val.to_string()));
-                    },
-                    "warning" | "rerun-if-changed" | "rerun-if-env-changed" => {
-                        bail!("`{}` is not supported in build script overrides", k);
-                    }
-                    _ => {
-                        let val = value.string(k)?.0;
-                        output.metadata.push((k.clone(), val.to_string()));
-                    }
-                }
-            }
-            ret.overrides.insert(lib_name, output);
-        }
-
-        Ok(ret)
     }
 }
