+++ src/cargo/core/compiler/context/mod.rs
@@ -97,10 +97,7 @@ impl<'a, 'cfg> Context<'a, 'cfg> {
         Ok(Self {
             bcx,
             compilation: Compilation::new(bcx),
-            build_state: Arc::new(BuildState::new(
-                &bcx.build_config.host,
-                &bcx.build_config.target,
-            )),
+            build_state: Arc::new(BuildState::new(&bcx.host_config, &bcx.target_config)),
             fingerprints: HashMap::new(),
             compiled: HashSet::new(),
             build_scripts: HashMap::new(),
